VEXIS-1/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent"
version = "1.0.0"
description = "Vision-based AI agent for GUI automation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AI Agent Team", email = "team@ai-agent.example"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Automation",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "automation", "gui", "vision", "screenshot"]
requires-python = ">=3.8"
dependencies = [
    "Pillow>=10.0.0",
    "pyautogui>=0.9.54",
    "requests>=2.31.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "pynput>=1.7.6",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "pyobjc-framework-Cocoa>=9.0; sys_platform=='darwin'",
    "pywin32>=306; sys_platform=='win32'",
    "python-xlib>=0.33; sys_platform=='linux'",
    "cryptography>=41.0.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[project.scripts]
ai-agent = "ai_agent.user_interface.main_app:main"
ai-agent-enhanced = "ai_agent.user_interface.enhanced_app:main"
ai-agent-validate = "ai_agent.user_interface.shell_wrapper:main"

[project.urls]
Homepage = "https://github.com/ai-agent/ai-agent"
Repository = "https://github.com/ai-agent/ai-agent"
Documentation = "https://ai-agent.readthedocs.io/"
"Bug Reports" = "https://github.com/ai-agent/ai-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ai_agent = [
    "config/*.yaml",
    "config/*.json", 
    "templates/*.txt",
    "scripts/*.sh",
    "scripts/*.ps1",
]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "cross_platform: marks tests as cross-platform tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
